Add "ordering" to classrooms, plus insert/update timestamps.
"Ordering" column has always been in the database but has always been ignored.  This update:
- sorts the Classrooms table by the ordering field;
- sets the ordering field to one plus the current max ordering, when a new classroom is submitted with no ordering specified;
- "makes room" for the ordering value of a new classroom, when it is submitted with a specified ordering, by incrementing the ordering field of all classooms whose ordering is equal to or greater than the value of the new classroom's ordering;
- probably some other stuff

Also adds insert and update timestamps to the classrooms table.  These fields are set on insert but are otherwise not used at present.  Insert timestamp works, update timestamp is correctly set on insert but needs to be updated when the record is updated.