Improved dialog styling.
Fixed the label/input alignment in Classroom dialog to match that of student dialog.
Changed styling of the focused input, removing (Bootstrap?) blue border and replaced with a green gradient background, matching the background of the buttons.  (Might actually not be a good idea, as there is now no way to indicate focus on buttons.)
 Removed all the vestigial CSS.