Add default value of null for each parameter to the LoginAttempt constructor.
Propel will typically create empty instances of Model classes when performing a find(), requiring a constructor without parameters.  However, it remains convenient to have a constructor WITH parameters when creating a new one to be inserted in the DB.  Providing a default value of null for each parameter allows for both of these situations.